# openapi.yaml
openapi: 3.1.0
info:
  title: Immigration Law API
  description: Analyze legal questions, draft motions, and summarize uploaded evidence for asylum and immigration proceedings.
  version: 1.0.0
servers:
  - url: https://api.lawqb.com
paths:
  /analyze:
    post:
      operationId: analyzeLegalQuestion
      summary: Analyze a legal question using IRAC format
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                jurisdiction:
                  type: string
                preferredSources:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: IRAC legal analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeResponse'
  /uploadEvidence:
    post:
      operationId: summarizeEvidenceUpload
      summary: Upload and summarize a document for asylum or immigration proceedings
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                jurisdiction:
                  type: string
                context:
                  type: string
      responses:
        '200':
          description: Evidence summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummarizeEvidenceResponse'
components:
  schemas:
    AnalyzeResponse:
      type: object
      properties:
        issue:
          type: string
        rule:
          type: string
        application:
          type: string
        conclusion:
          type: string
        citations:
          type: array
          items:
            type: string
        conflictsOrAmbiguities:
          type: string
        verificationNotes:
          type: string
  SummarizeEvidenceResponse:
    type: object
    properties:
      filename:
        type: string
      sizeInBytes:
        type: integer
      readableSize:
        type: string
      fileType:
        type: string
      truncated:
        type: boolean
      category:
        type: string
        description: "Classification of the document ('case facts' or 'country conditions')"
      summary:
        type: string
      keyFacts:
        type: array
        items:
          type: string
      legalIssues:
        type: array
        items:
          type: string
      credibilityConcerns:
        type: string
      recommendation:
        type: string
      verificationNotes:
        type: string